move file into separate directories
